requirements.txt
fastapi[all]
google-auth
google-auth-oauthlib
python-dotenv
uvicorn
requests
starlette